fix(zero-cache): speed up initial sync tests
