+++ packages/zero-cache/src/services/replicator/initial-sync.test.ts
@@ -370,7 +370,7 @@ describe('replicator/initial-sync', () => {
     await dropReplicationSlot(upstream, replicationSlot(REPLICA_ID));
     await testDBs.drop(upstream);
     await replicaFile.unlink();
-  }, 10000);
+  });
 
   for (const c of cases) {
     test(`startInitialDataSynchronization: ${c.name}`, async () => {
@@ -429,92 +429,6 @@ describe('replicator/initial-sync', () => {
         slotName: replicationSlot(REPLICA_ID),
         lsn: replicaState.watermark,
       });
-    }, 10000);
-
-    type InvalidUpstreamCase = {
-      error: string;
-      setupUpstreamQuery?: string;
-      upstream?: Record<string, object[]>;
-    };
-
-    const invalidUpstreamCases: InvalidUpstreamCase[] = [
-      {
-        error: 'does not have a PRIMARY KEY',
-        setupUpstreamQuery: `
-        CREATE TABLE issues("issueID" INTEGER, "orgID" INTEGER);
-      `,
-      },
-      {
-        error: 'uses reserved column name "_0_version"',
-        setupUpstreamQuery: `
-        CREATE TABLE issues(
-          "issueID" INTEGER PRIMARY KEY, 
-          "orgID" INTEGER, 
-          _0_version INTEGER);
-      `,
-      },
-      {
-        error: 'Schema "_zero" is reserved for internal use',
-        setupUpstreamQuery: `
-        CREATE SCHEMA _zero;
-        CREATE TABLE _zero.is_not_allowed(
-          "issueID" INTEGER PRIMARY KEY, 
-          "orgID" INTEGER
-        );
-        CREATE PUBLICATION zero_foo FOR TABLES IN SCHEMA _zero;
-        `,
-      },
-      {
-        error: 'Only the default "public" schema is supported',
-        setupUpstreamQuery: `
-        CREATE SCHEMA unsupported;
-        CREATE TABLE unsupported.issues ("issueID" INTEGER PRIMARY KEY, "orgID" INTEGER);
-        CREATE PUBLICATION zero_foo FOR TABLES IN SCHEMA unsupported;
-      `,
-      },
-      {
-        error: 'Table "table/with/slashes" has invalid characters',
-        setupUpstreamQuery: `
-        CREATE TABLE "table/with/slashes" ("issueID" INTEGER PRIMARY KEY, "orgID" INTEGER);
-      `,
-      },
-      {
-        error: 'Table "table.with.dots" has invalid characters',
-        setupUpstreamQuery: `
-        CREATE TABLE "table.with.dots" ("issueID" INTEGER PRIMARY KEY, "orgID" INTEGER);
-      `,
-      },
-      {
-        error:
-          'Column "column/with/slashes" in table "issues" has invalid characters',
-        setupUpstreamQuery: `
-        CREATE TABLE issues ("issueID" INTEGER PRIMARY KEY, "column/with/slashes" INTEGER);
-      `,
-      },
-      {
-        error:
-          'Column "column.with.dots" in table "issues" has invalid characters',
-        setupUpstreamQuery: `
-        CREATE TABLE issues ("issueID" INTEGER PRIMARY KEY, "column.with.dots" INTEGER);
-      `,
-      },
-    ];
-
-    for (const c of invalidUpstreamCases) {
-      test(`Invalid upstream: ${c.error}`, async () => {
-        await initDB(upstream, c.setupUpstreamQuery, c.upstream);
-
-        const result = await initialSync(
-          createSilentLogContext(),
-          REPLICA_ID,
-          replica,
-          upstream,
-          getConnectionURI(upstream, 'external'),
-        ).catch(e => e);
-
-        expect(result).toBeInstanceOf(Error);
-        expect(String(result)).toContain(c.error);
-      });
-    }
+    });
   }
 });
